requirements.txt
fastapi 
uvicorn 
python-dotenv 
transformers
torch 
torchvision 
torchaudio
python-multipart
httpx
llama-cpp-python
huggingface_hub